Moved things around to reflect this style of organization better; consolidated stages under stages/, made the core stage management code be in stage.py, root directory for submodules and core functionality only.
