+++ pyproject.toml
@@ -34,19 +34,17 @@ docs = [
 ]
 
 [project.scripts]
-## General
-MEDS_transform-aggregate_code_metadata = "MEDS_transforms.aggregate_code_metadata:main"
-MEDS_transform-fit_vocabulary_indices = "MEDS_transforms.fit_vocabulary_indices:main"
-MEDS_transform-reshard_to_split = "MEDS_transforms.reshard_to_split:main"
-## Filters
-MEDS_transform-filter_measurements = "MEDS_transforms.filters.filter_measurements:main"
-MEDS_transform-filter_subjects = "MEDS_transforms.filters.filter_subjects:main"
-## Transforms
-MEDS_transform-reorder_measurements = "MEDS_transforms.transforms.reorder_measurements:main"
-MEDS_transform-add_time_derived_measurements = "MEDS_transforms.transforms.add_time_derived_measurements:main"
-MEDS_transform-extract_values = "MEDS_transforms.transforms.extract_values:main"
-MEDS_transform-normalization = "MEDS_transforms.transforms.normalization:main"
-MEDS_transform-occlude_outliers = "MEDS_transforms.transforms.occlude_outliers:main"
+# Stages
+MEDS_transform-aggregate_code_metadata = "MEDS_transforms.stages.aggregate_code_metadata:main"
+MEDS_transform-fit_vocabulary_indices = "MEDS_transforms.stages.fit_vocabulary_indices:main"
+MEDS_transform-reshard_to_split = "MEDS_transforms.stages.reshard_to_split:main"
+MEDS_transform-filter_measurements = "MEDS_transforms.stages.filter_measurements:main"
+MEDS_transform-filter_subjects = "MEDS_transforms.stages.filter_subjects:main"
+MEDS_transform-reorder_measurements = "MEDS_transforms.stages.reorder_measurements:main"
+MEDS_transform-add_time_derived_measurements = "MEDS_transforms.stages.add_time_derived_measurements:main"
+MEDS_transform-extract_values = "MEDS_transforms.stages.extract_values:main"
+MEDS_transform-normalization = "MEDS_transforms.stages.normalization:main"
+MEDS_transform-occlude_outliers = "MEDS_transforms.stages.occlude_outliers:main"
 
 # Runner
 MEDS_transform-runner = "MEDS_transforms.runner:main"
