--- src/MEDS_transforms/filters/README.md
@@ -1,5 +1,0 @@
-# Filters
-
-Filters remove wholesale events within the data, either at the subject or event level. For transformations
-that simply _occlude_ aspects of the data (e.g., by setting a code variable to `UNK`), see the `transforms`
-library section.
