+++ src/MEDS_transforms/stages/add_time_derived_measurements.py
@@ -6,9 +6,8 @@ from collections.abc import Callable
 import polars as pl
 from omegaconf import DictConfig, OmegaConf
 
-from MEDS_transforms import INFERRED_STAGE_KEYS
-
-from ..stages import MEDS_transforms_stage
+from .. import INFERRED_STAGE_KEYS
+from ..stage import MEDS_transforms_stage
 
 logger = logging.getLogger(__name__)
