+++ src/MEDS_transforms/stages/reshard_to_split.py
@@ -14,10 +14,9 @@ import polars as pl
 from meds import subject_id_field, subject_splits_filepath, time_field
 from omegaconf import DictConfig, OmegaConf
 
-from MEDS_transforms.mapreduce import rwlock_wrap, shard_iterator, shuffle_shards
-from MEDS_transforms.utils import stage_init, write_lazyframe
-
-from .stages import MEDS_transforms_stage
+from ..mapreduce import rwlock_wrap, shard_iterator, shuffle_shards
+from ..stage import MEDS_transforms_stage
+from ..utils import stage_init, write_lazyframe
 
 logger = logging.getLogger(__name__)
